Split up the CSS
