+++ todoqueue_frontend/src/App.css
@@ -10,3 +10,4 @@ h1 {
     color: #333;
     text-align: left;
 }
+
