+++ todoqueue_frontend/src/components/popups/CompleteTaskPopup.js
@@ -1,6 +1,7 @@
 import React, { useState } from 'react';
 import BasePopup from './BasePopup';
 import { createWorkLog } from '../../api/tasks';
+import './popups.css';
 
 const CompleteTaskPopup = React.forwardRef((props, ref) => {
     const [completionUsers, setCompletionUsers] = useState([]);
