+++ todoqueue_frontend/src/components/popups/popups.css
@@ -1,0 +1,63 @@
+.input-field {
+    padding: 5px;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+}
+
+.form-section {
+    margin-bottom: 1rem;
+}
+
+.form-input {
+    padding: 0.5rem;
+    margin-top: 0.2rem;
+}
+
+.label-and-input {
+    display: flex;
+    align-items: center;
+}
+
+.label-and-input label {
+    margin-right: 0.5rem;
+    margin-left: 0.5rem;
+}
+
+.label-and-input span {
+    margin-left: 0.5rem;
+}
+
+.task-form {
+    display: flex;
+    flex-direction: column;
+}
+
+.error-message {
+    color: red;
+    margin-top: 20px;
+}
+
+.popup {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background-color: rgba(0, 0, 0, 0.5);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    z-index: 10;
+}
+
+.popup-inner {
+    background: white;
+    color: black;
+    padding: 24px;
+    border-radius: 12px;
+    width: 600px;
+    text-align: center;
+    box-shadow: 0 4px 16px rgba(255, 239, 0, 0.51);
+    /* 0.5s transition for background color, color, and appearing/disappearing */
+    transition: background-color 1s ease, opacity 0.5s ease, color 0.5s ease;
+}
